// apps/dashboard/src/api/model/dtoCreateInvoiceItemRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Alizzah Manajemen API
 * API for Alizzah School Management System
 * OpenAPI spec version: 1.0
 */

export interface DtoCreateInvoiceItemRequest {
	amount?: number;
	/** @maxLength 30 */
	category: string;
	/** @maxLength 100 */
	name: string;
	notes?: string;
	quantity?: number;
	unit_price?: number;
}