// apps/dashboard/src/api/model/dtoCreateInvoiceRequestType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Alizzah Manajemen API
 * API for Alizzah School Management System
 * OpenAPI spec version: 1.0
 */

export type DtoCreateInvoiceRequestType =
	(typeof DtoCreateInvoiceRequestType)[keyof typeof DtoCreateInvoiceRequestType];

export const DtoCreateInvoiceRequestType = {
	arrears: "arrears",
	manual: "manual",
} as const;